+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # (MODX)EvolutionCMS.snippets.ddRunSnippets changelog
 
 
+## Version 2.0 (2012-03-10)
+* \+ Snippet redeveloped to run few snippets.
+* \* Large amount of code changes.
+* \* Params `name`, `parameters` and `values` renamed into `snipName`, `snipParams` and `snipValues` respectively.
+
+
 ## Version 1.0 (2011-12-12)
 * \+ The first release.